SmartApp - IoT App for Android
In a world where a user has appliances that connect to a home network,
he can chose when his appliances can run and how long they can run. This
allows all his appliances to be automated without him having to manually
take care of switching them on or off.
He can then let his Home server schedule based on when the price for
energy per hour, supplied by the power company is low. He can then send
this schedule to the power company which can then see whether there is
any violation to existing global schedule of all user. If there is, the
Utility server will schedule all the appliances for the user and the
user is presented with options of whether he wants to accept it or stick
with his initial scheduling.